Allow all approves
The approve function should not revert if it approves a non 0 value, even if it's already non 0. The ERC20 standard states that the interface should prevent it, not the contract.
"NOTE: To prevent attack vectors like the one described here and discussed here, clients SHOULD make sure to create user interfaces in such a way that they set the allowance first to 0 before setting it to another value for the same spender. THOUGH The contract itself shouldn't enforce it, to allow backwards compatibility with contracts deployed before"
https://github.com/ethereum/EIPs/blob/master/EIPS/eip-20-token-standard.md

This PR follows this issue: https://github.com/Giveth/minime/issues/45